Solved some of the problems for the complete amount of points.

+++ Inheritance/Restaurant/Product.cs
@@ -7,12 +7,12 @@ namespace Restaurant
     class Product
     {
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public Product(string name, decimal price)
         {
             Name = name;
-            Price = (double)price;
+            Price = price;
         }
     }
 }
